+++ HiveWE/Texture.cpp
@@ -20,7 +20,7 @@ Texture::Texture(const fs::path& path) {
 			std::swap(minimap_color.r, minimap_color.b);
 		}
 	} else {
-		uint8_t* image_data = SOIL_load_image_from_memory(reader.buffer.data(), reader.buffer.size(), &width, &height, &channels, SOIL_LOAD_AUTO);
+		uint8_t* image_data = SOIL_load_image_from_memory(reader.buffer.data(), static_cast<int>(reader.buffer.size()), &width, &height, &channels, SOIL_LOAD_AUTO);
 		data = std::vector<uint8_t>(image_data, image_data + width * height * channels);
 		delete image_data;
 	}
